+++ experimental/packages/configuration/scripts/generate-config.js
@@ -25,6 +25,11 @@ const Ajv = require('ajv/dist/2020');
 const standaloneCode = require('ajv/dist/standalone').default;
 const typescript = require('typescript');
 
+// Get latest version by running:
+//    git tag -l --sort=version:refname | grep -v -- - | tail -1
+// in git@github.com:open-telemetry/opentelemetry-configuration.git
+const CONFIG_VERSION = 'v1.0.0';
+
 const TOP = path.resolve(__dirname, '..');
 const SCHEMA_PATH = path.join(
   TOP,
@@ -105,10 +110,13 @@ function removeDuplicateTsDeclarations(tsCode) {
 // ---- 1. Get and load the OpenTelemetry Configuration JSON schema.
 
 // Run the bash script to clone / refresh the schema repository
-const bashResult = spawnSync('bash', ['generate-config.sh'], {
-  stdio: 'inherit',
-  cwd: __dirname,
-});
+const bashResult = spawnSync(
+  'bash',
+  ['clone-config-repo-at-tag.sh', CONFIG_VERSION],
+  {
+    stdio: 'inherit',
+    cwd: __dirname,
+  });
 if (bashResult.error) {
   throw bashResult.error;
 }
@@ -127,7 +135,8 @@ const validatorJs = standaloneCode(ajvAot, validateFn);
 const validatorJsWithHeader = [
   '// AUTO-GENERATED — do not edit',
   '// Pre-compiled ajv validator for the OpenTelemetry configuration schema',
-  '// Run `npm run generate:config` from the configuration package to regenerate',
+  `// Generated from opentelemetry-configuration.git ${CONFIG_VERSION}`,
+  '// Run `npm run generate:config` to regenerate',
   '// eslint-disable-next-line',
   '',
   validatorJs,
@@ -144,7 +153,8 @@ const validatorDts = [
   '/* eslint-disable */',
   '// AUTO-GENERATED — do not edit',
   '// Pre-compiled ajv validator for the OpenTelemetry configuration schema',
-  '// Run `npm run generate:config` from the configuration package to regenerate',
+  `// Generated from opentelemetry-configuration.git ${CONFIG_VERSION}`,
+  '// Run `npm run generate:config` to regenerate',
   '',
   '/** Minimal subset of ajv ErrorObject used by FileConfigFactory */',
   'interface ValidatorError {',
@@ -171,8 +181,8 @@ const bannerComment = [
   '',
   '//',
   '// AUTO-GENERATED — do not edit',
-  '// Generated from opentelemetry-configuration JSON schema v1.0.0',
-  '// Run `npm run generate:config` from the configuration package to regenerate',
+  `// Generated from opentelemetry-configuration.git ${CONFIG_VERSION}`,
+  '// Run `npm run generate:config` to regenerate',
   '//',
   '/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-empty-object-type */',
 ].join('\n');